Prevents NMM from removing screenshot files from newly added fomods. Also fixes NMM crashing when the screenshot is missing.

+++ FOMod/FOMod.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Drawing;
 using System.IO;
 using System.Text.RegularExpressions;
 using System.Xml;
@@ -732,7 +733,13 @@ namespace Nexus.Client.Mods.Formats.FOMod
 		public byte[] GetFile(string p_strFile)
 		{
 			if (!ContainsFile(p_strFile))
-				throw new FileNotFoundException("File doesn't exist in FOMod", p_strFile);
+			{
+				if (Path.GetFileNameWithoutExtension(p_strFile).ToLower() == "screenshot")
+					return (byte[])(new ImageConverter().ConvertTo(new Bitmap(1, 1), typeof(byte[])));
+				else
+					throw new FileNotFoundException("File doesn't exist in FOMod", p_strFile);
+			}		
+				
 			if ((m_arcCacheFile != null) && m_arcCacheFile.ContainsFile(GetRealPath(p_strFile)))
 				return m_arcCacheFile.GetFileContents(GetRealPath(p_strFile));
 			return m_arcFile.GetFileContents(GetRealPath(p_strFile));
@@ -878,7 +885,7 @@ namespace Nexus.Client.Mods.Formats.FOMod
 			{
 				if (p_mifInfo.Screenshot == null)
 				{
-					if (Screenshot != null)
+					if ((Screenshot != null) && p_booOverwriteAllValues)
 					{
 						DeleteFile(m_strScreenshotPath);
 						Screenshot = p_mifInfo.Screenshot;
